Refactor `AUTH_RESPONSE` Import Path
