+++ src/modules/collection/interface/__test__/collection.e2e.spec.ts
@@ -12,12 +12,11 @@ import { loadFixtures } from '@data/util/loader';
 import { AppModule } from '@/app.module';
 import { COGNITO_AUTH } from '@/common/cognito/application/interface/cognito.service.interface';
 import { SuccessResponseInterceptor } from '@/common/response_service/interceptor/success_response.interceptor';
-import { AUTH_RESPONSE } from '@/modules/authorization/infraestructure/policy/exceptions/auth-error';
+import { AUTH_RESPONSE } from '@/modules/auth/application/exceptions/auth-error';
 import { ENVIROMENT_RESPONSE } from '@/modules/enviroment/application/exceptions/enviroment-response.enum';
 import { TEAM_RESPONSE } from '@/modules/team/application/exceptions/team-response.enum';
 import { identityProviderServiceMock } from '@/test/test.module.bootstrapper';
 import { DataObject, createAccessToken, makeRequest } from '@/test/test.util';
-
 
 import { COLLECTION_RESPONSE } from '../../application/exceptions/collection-response.enum';
